+++ index.js
@@ -18,6 +18,7 @@ var request = require('superagent');
 var logger = require('./services/logger');
 var Integrator = require('./integrations');
 var errorHandler = require('./services/error-handler');
+var codeSyntaxInspector = require('./utils/code-syntax-inspector');
 
 function requireAllModels(Implementation, modelsDir, displayMessage) {
   if (modelsDir) {
@@ -114,10 +115,18 @@ exports.init = function (Implementation) {
             directorySmartImplementation = path.resolve('.') + '/forest';
           }
 
-          // NOTICE: Do not display an error log if the forest/ directory does
-          //         not exist.
-          return requireAllModels(Implementation, directorySmartImplementation,
-            false);
+          return codeSyntaxInspector
+            .extractCodeSyntaxErrorInDirectoryFile(directorySmartImplementation)
+            .then(function (hasError) {
+              if (hasError) {
+                throw new Error();
+              } else {
+                // NOTICE: Do not display an error log if the forest/ directory
+                //         does not exist.
+                return requireAllModels(Implementation,
+                  directorySmartImplementation, false);
+              }
+            });
         })
         .thenReturn(models);
     })
@@ -217,6 +226,11 @@ exports.init = function (Implementation) {
               });
         }
       }
+    })
+    .catch(function () {
+      logger.error('An error occured while computing the Forest apimap. Your ' +
+        'application apimap cannot be sent to Forest. Your Admin UI might ' +
+        'not reflect your application models.');
     });
 
   if (opts.expressParentApp) {