+++ utils/code-syntax-inspector.js
@@ -1,0 +1,34 @@
+'use strict';
+var P = require('bluebird');
+var esprima = require('esprima');
+var path = require('path');
+var fs = require('fs');
+var Pfs = P.promisifyAll(fs);
+var logger = require('../services/logger');
+
+exports.extractCodeSyntaxErrorInDirectoryFile = function (modelsDir) {
+  var hasError = false;
+
+  if (fs.existsSync(modelsDir)) {
+    return Pfs.readdirAsync(modelsDir)
+      .map(function (filename) {
+        var file = path.join(modelsDir, filename);
+        var fileContent = Pfs.readFileSync(file);
+
+        try {
+          return esprima.parse(fileContent.toString(), { loc: true });
+        } catch (error) {
+          if (error) {
+            hasError = true;
+            logger.error('Forest customization failed due to a syntax error: ' +
+              error.description + ' in ' + file + ':' + error.lineNumber);
+          }
+        }
+      })
+      .then(function () { return hasError; });
+  } else {
+    return new P(function(resolve) {
+      return resolve(hasError);
+    });
+  }
+};